+++ components/entities/api/src/main/java/org/phenotips/entities/internal/AbstractPrimaryEntityGroupManager.java
@@ -17,7 +17,6 @@
  */
 package org.phenotips.entities.internal;
 
-import org.phenotips.Constants;
 import org.phenotips.components.ComponentManagerRegistry;
 import org.phenotips.entities.PrimaryEntity;
 import org.phenotips.entities.PrimaryEntityGroupManager;
@@ -26,19 +25,12 @@ import org.phenotips.entities.PrimaryEntityManager;
 import org.xwiki.bridge.DocumentAccessBridge;
 import org.xwiki.component.manager.ComponentLookupException;
 import org.xwiki.component.manager.ComponentManager;
-import org.xwiki.model.EntityType;
 import org.xwiki.model.reference.EntityReference;
 import org.xwiki.model.reference.EntityReferenceSerializer;
-import org.xwiki.query.Query;
-import org.xwiki.query.QueryException;
 import org.xwiki.query.QueryManager;
-import org.xwiki.stability.Unstable;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.LinkedList;
-import java.util.List;
 
 import javax.inject.Provider;
 
@@ -48,29 +40,11 @@ import org.slf4j.LoggerFactory;
 
 import com.xpn.xwiki.XWikiContext;
 import com.xpn.xwiki.doc.XWikiDocument;
-import com.xpn.xwiki.objects.BaseObject;
 
 /**
- * Base class for implementing specific entity groups, where members declare the groups that they belong to.
+ * Base class for implementing specific entity groups, where members declare the groups that they belong to. The class
+ * deals with logical group operations and not with implementation.
  * <p>
- * By default, this uses the document name as the identifier, the document title as the name, and either a
- * {@code description} property in the main XObject, or the document content, as the description. If two objects use the
- * same document for storage, they are assumed to be equal, and no actual data equality is checked.
- * </p>
- * <p>
- * Members declare their group membership by attaching an XObject of type {@link #GROUP_MEMBERSHIP_CLASS} pointing to
- * the group document.
- * </p>
- * <p>
- * In order to function properly, a {@link PrimaryEntityManager} component with the hint ({@code @Named}) set to the
- * name of {@link #getMemberType() the XClass used for members} must exist. Shorter versions of the XClass name are also
- * accepted. For example, for members of type {@code PhenoTips.PatientClass}, the following names are supported:
- * </p>
- * <ol>
- * <li>PhenoTips.PatientClass</li>
- * <li>PatientClass</li>
- * <li>Patient</li>
- * </ol>
  *
  * @param <G> the type of the group containing the entities
  * @param <E> the type of entities belonging to this group; if more than one type of entities can be part of the group,
@@ -78,27 +52,27 @@ import com.xpn.xwiki.objects.BaseObject;
  * @version $Id$
  * @since 1.3M2
  */
-@Unstable("New class and interface added in 1.3")
-public abstract class AbstractPrimaryEntityGroup<G extends PrimaryEntity, E extends PrimaryEntity>
+public abstract class AbstractPrimaryEntityGroupManager<G extends PrimaryEntity, E extends PrimaryEntity>
     implements PrimaryEntityGroupManager<G, E>
 {
-    /** The XClass used for storing membership information by default. */
-    protected static final EntityReference GROUP_MEMBERSHIP_CLASS =
-        new EntityReference("EntityBindingClass", EntityType.DOCUMENT, Constants.CODE_SPACE_REFERENCE);
-
+    /** Entity type of group object. */
     protected EntityReference groupEntityReference;
 
+    /** Manager of group primary entities. */
+    protected final PrimaryEntityManager<G> groupManager;
+
+    /** Entity type of member object. */
     protected EntityReference memberEntityReference;
 
+    /** Manager of member primary entities. */
     protected final PrimaryEntityManager<E> membersManager;
-
-    protected final PrimaryEntityManager<G> groupManager;
 
     /** Logging helper object. */
     protected final Logger logger = LoggerFactory.getLogger(getClass());
 
     @SuppressWarnings("unchecked")
-    protected AbstractPrimaryEntityGroup(EntityReference groupEntityReference, EntityReference memberEntityReference)
+    protected AbstractPrimaryEntityGroupManager(
+        EntityReference groupEntityReference, EntityReference memberEntityReference)
     {
         this.groupEntityReference = groupEntityReference;
         this.memberEntityReference = memberEntityReference;
@@ -138,41 +112,6 @@ public abstract class AbstractPrimaryEntityGroup<G extends PrimaryEntity, E exte
     public Collection<E> getMembers(G group)
     {
         return getMembersOfType(group, memberEntityReference);
-    }
-
-    @Override
-    public Collection<E> getMembersOfType(G group, EntityReference type)
-    {
-        Collection<E> result = new LinkedList<>();
-        try {
-            StringBuilder hql = new StringBuilder();
-            hql.append("select distinct binding.name from BaseObject binding, StringProperty groupReference");
-            if (type != null) {
-                hql.append(", BaseObject entity");
-            }
-            hql.append(" where binding.className = :memberClass")
-                .append(" and groupReference.id.id = binding.id and groupReference.id.name = :referenceProperty")
-                .append(" and groupReference.value = :selfReference");
-            if (type != null) {
-                hql.append(" and entity.name = binding.name and entity.className = :entityType");
-            }
-
-            Query q = getQueryManager().createQuery(hql.toString(), Query.HQL);
-
-            q.bindValue("memberClass", getLocalSerializer().serialize(GROUP_MEMBERSHIP_CLASS));
-            q.bindValue("referenceProperty", getMembershipProperty());
-            q.bindValue("selfReference", getFullSerializer().serialize(group.getDocumentReference()));
-            if (type != null) {
-                q.bindValue("entityType", getLocalSerializer().serialize(type));
-            }
-            List<String> memberIds = q.execute();
-            for (String memberId : memberIds) {
-                result.add(this.membersManager.get(memberId));
-            }
-        } catch (QueryException ex) {
-            this.logger.warn("Failed to query members: {}", ex.getMessage());
-        }
-        return result;
     }
 
     @Override
@@ -238,92 +177,6 @@ public abstract class AbstractPrimaryEntityGroup<G extends PrimaryEntity, E exte
         return success;
     }
 
-    @Override
-    public boolean addMember(G group, E member)
-    {
-        try {
-            DocumentAccessBridge dab =
-                ComponentManagerRegistry.getContextComponentManager().getInstance(DocumentAccessBridge.class);
-            XWikiDocument doc = (XWikiDocument) dab.getDocument(member.getDocumentReference());
-            BaseObject obj = doc.getXObject(GROUP_MEMBERSHIP_CLASS, getMembershipProperty(),
-                getFullSerializer().serialize(group.getDocumentReference()), false);
-            if (obj != null) {
-                return true;
-            }
-            obj = doc.newXObject(GROUP_MEMBERSHIP_CLASS, getXContext());
-            obj.setStringValue(getMembershipProperty(), getFullSerializer().serialize(group.getDocumentReference()));
-            this.setMemberParameters(member, obj);
-            getXContext().getWiki().saveDocument(doc, "Added to group " + group.getDocumentReference(), true, getXContext());
-            return true;
-        } catch (Exception ex) {
-            this.logger.warn("Failed to add member to group: {}", ex.getMessage());
-        }
-        return false;
-    }
-
-    /**
-     * If an extending class wants to record parameters with a member, override this.
-     *
-     * @param member that is saved
-     * @param obj object for recording parameters
-     */
-    protected void setMemberParameters(E member, BaseObject obj)
-    {
-        // do nothing as default.
-    }
-
-    @Override
-    public boolean removeMember(G group, E member)
-    {
-        try {
-            XWikiDocument doc = getXWikiDocument(member);
-            BaseObject obj = doc.getXObject(GROUP_MEMBERSHIP_CLASS, getMembershipProperty(),
-                getFullSerializer().serialize(group.getDocumentReference()), false);
-            if (obj == null) {
-                return true;
-            }
-            doc.removeXObject(obj);
-            getXContext().getWiki().saveDocument(doc, "Removed from group " + group.getDocumentReference(), true, getXContext());
-            return true;
-        } catch (Exception ex) {
-            this.logger.warn("Failed to remove member from group: {}", ex.getMessage());
-        }
-        return false;
-    }
-
-    @Override
-    public Collection<G> getGroupsForMember(PrimaryEntity member)
-    {
-        try {
-            // FIXME GROUP_MEMBERSHIP_CLASS should be replaced with a static method call
-            String bindingClassName = this.getLocalSerializer().serialize(GROUP_MEMBERSHIP_CLASS);
-            String xwikiId = this.getXContext().getWikiId() + ":";
-            String groupClass = this.getLocalSerializer().serialize(groupEntityReference);
-            Query q = this.getQueryManager().createQuery(
-                ", BaseObject obj, BaseObject groupObj, StringProperty property "
-                + "where obj.id.id = property.id and "
-                + "property.value = concat('" + xwikiId + "', doc.fullName) and "
-                + "property.id.name = 'reference' and "
-                + "obj.className = '" + bindingClassName + "' and "
-                + "obj.name = :name and "
-                + "doc.space = :gspace and "
-                + "groupObj.name = doc.fullName and "
-                + "groupObj.className = '" + groupClass + "'", Query.HQL);
-            q.bindValue("gspace", this.groupManager.getDataSpace().getName());
-            q.bindValue("name", this.getLocalSerializer().serialize(member.getDocument()));
-            List<String> docNames = q.execute();
-            Collection<G> result = new ArrayList<>(docNames.size());
-            for (String docName : docNames) {
-                result.add(this.groupManager.get(docName));
-            }
-            return result;
-        } catch (QueryException ex) {
-            this.logger.warn("Failed to query all entities of type [{}]: {}", groupEntityReference,
-                ex.getMessage());
-        }
-        return Collections.emptyList();
-    }
-
     protected QueryManager getQueryManager()
     {
         try {
@@ -383,10 +236,5 @@ public abstract class AbstractPrimaryEntityGroup<G extends PrimaryEntity, E exte
         DocumentAccessBridge dab = getDataAccessBridge();
         XWikiDocument doc = (XWikiDocument) dab.getDocument(p.getDocument());
         return doc;
-    }
-
-    protected String getMembershipProperty()
-    {
-        return REFERENCE_XPROPERTY;
     }
 }
